requirements.txt
Flask
Flask-Session
psycopg2-binary
SQLAlchemy
sqlalchemy.orm
werkzeug.exceptions
werkzeug.security
functools
